+++ src/collective/z3cform/jsonwidget/browser/static/js/fields/ReferenceField.js
@@ -6,10 +6,11 @@ import Modal from 'react-modal';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faChevronRight,
-  faPlus,
   faHome,
   faTimes,
+  faTrash,
 } from '@fortawesome/free-solid-svg-icons';
+
 import './ReferenceField.less';
 
 const Breadcrumbs = ({ fetchData, breadcrumbs }) => (
@@ -51,6 +52,31 @@ const Breadcrumbs = ({ fetchData, breadcrumbs }) => (
     ))}
   </div>
 );
+
+const ItemElement = ({ isSelected, result, onAddReference }) => {
+  const { getTranslationFor } = useContext(WidgetContext);
+
+  if (isSelected) {
+    return (
+      <span className={`content-title ${isSelected ? 'selected-item' : ''}`}>
+        {result.Title}
+      </span>
+    );
+  }
+  return (
+    <a
+      href="#"
+      className={`content-title ${isSelected ? 'selected-item' : ''}`}
+      title={getTranslationFor('Add') + ' ' + result.Title}
+      onClick={e => {
+        e.preventDefault();
+        onAddReference(result);
+      }}
+    >
+      {result.Title}
+    </a>
+  );
+};
 
 const ReferenceField = ({ value, id, row, items }) => {
   const { root } = items;
@@ -164,7 +190,7 @@ const ReferenceField = ({ value, id, row, items }) => {
                 }}
                 title={getTranslationFor('Delete')}
               >
-                <FontAwesomeIcon icon={faTimes} />
+                <FontAwesomeIcon icon={faTrash} />
               </button>
               <a href={ref.getURL} target="_blank" rel="noopener noreferrer">
                 {ref.Title}
@@ -189,6 +215,9 @@ const ReferenceField = ({ value, id, row, items }) => {
         style={customStyles}
         ariaHideApp={false}
       >
+        <button className="close" onClick={closeModal}>
+          <FontAwesomeIcon icon={faTimes} />
+        </button>
         <div className="modal-content-wrapper">
           <h2>{getTranslationFor('Select contents')}</h2>
           <p className="discreet">
@@ -196,7 +225,6 @@ const ReferenceField = ({ value, id, row, items }) => {
               'Navigate through site structure and select one or more contents.',
             )}
           </p>
-          <button onClick={closeModal}>{getTranslationFor('Close')}</button>
 
           <Breadcrumbs
             fetchData={fetchData}
@@ -205,24 +233,11 @@ const ReferenceField = ({ value, id, row, items }) => {
           <div className="content-results-wrapper">
             {modalData.results.map(result => (
               <div className="content-item" key={result.UID}>
-                <button
-                  type="button"
-                  disabled={selectedUIDs.includes(result.UID)}
-                  onClick={e => {
-                    e.preventDefault();
-                    onAddReference(result);
-                  }}
-                  title={getTranslationFor('Add')}
-                >
-                  <FontAwesomeIcon icon={faPlus} />
-                </button>
-                <span
-                  className={`content-title ${
-                    selectedUIDs.includes(result.UID) ? 'selected-item' : ''
-                  }`}
-                >
-                  {result.Title}
-                </span>
+                <ItemElement
+                  isSelected={selectedUIDs.includes(result.UID)}
+                  result={result}
+                  onAddReference={onAddReference}
+                />
                 {result.is_folderish && (
                   <button
                     type="button"
@@ -274,6 +289,12 @@ ReferenceField.propTypes = {
   value: PropTypes.array,
   id: PropTypes.string,
   row: PropTypes.number,
+};
+
+ItemElement.propTypes = {
+  isSelected: PropTypes.bool,
+  result: PropTypes.object,
+  onAddReference: PropTypes.func,
 };
 
 export default ReferenceField;